more automated testing and pipe,procfs functionality

+++ src/process.cpp
@@ -1,13 +1,25 @@
 #include <libtbd/process.hpp>
 #include <libtbd/error.hpp>
+#include <libtbd/pipe.hpp>
 #include <sys/ptrace.h>
 #include <iostream>
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <unistd.h>
 
+namespace {
+	void exit_with_perror(
+			tbd::pipe& channel, std::string const& prefix) {
+		auto message = prefix + ": " + std::strerror(errno);
+		channel.write(
+				reinterpret_cast<std::byte*>(message.data()), message.size());
+		exit(-1);
+	}
+}
 
-std::unique_ptr<tbd::process> tbd::process::launch(std::filesystem::path path) {
+
+std::unique_ptr<tbd::process> tbd::process::launch(std::filesystem::path path, bool debug) {
+	pipe channel(/*close_on_exec=*/true);
 	pid_t pid;
 
 	if ((pid = fork()) < 0) {
@@ -16,19 +28,32 @@ std::unique_ptr<tbd::process> tbd::process::launch(std::filesystem::path path) {
 	}
 
 	if (pid == 0) {
-		if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
+		if (debug and ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
 			// Error: Tracing failed
-			error::send_errno("Tracing failed");
+			exit_with_perror(channel, "Tracing failed");
 		}
 		if (execlp(path.c_str(), path.c_str(), nullptr) < 0) {
 			// Error: exec failed
-			error::send_errno("exec failed");
+			exit_with_perror(channel, "exec failed");
 		}
 	}
 
-	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end=*/true));
+	channel.close_write();
+	auto data = channel.read();
+	channel.close_read();
 
-	proc->wait_on_signal();
+	if (data.size() > 0) {
+		waitpid(pid, nullptr, 0);
+		auto chars = reinterpret_cast<char*>(data.data());
+		error::send(std::string(chars, chars + data.size()));
+	}
+
+
+	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end=*/true, debug));
+
+	if (debug) {
+		proc->wait_on_signal();
+	}
 
 	return proc;
 }
@@ -43,7 +68,7 @@ std::unique_ptr<tbd::process> tbd::process::attach(pid_t pid) {
 		error::send_errno("Could not attach");
 	}
 
-	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end=*/false));
+	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end=*/false, /*attached=*/true));
 	proc->wait_on_signal();
 
 	return proc;
@@ -86,12 +111,14 @@ tbd::stop_reason tbd::process::wait_on_signal() {
 tbd::process::~process() {
 	if (pid_ != 0) {
 		int status;
-		if (state_ == process_state::running) {
-			kill(pid_, SIGSTOP);
-			waitpid(pid_, &status, 0);
+		if (is_attached_) {
+			if (state_ == process_state::running) {
+				kill(pid_, SIGSTOP);
+				waitpid(pid_, &status, 0);
+			}
+			ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+			kill(pid_, SIGCONT);
 		}
-		ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-		kill(pid_, SIGCONT);
 
 		if (terminate_on_end_) {
 			kill(pid_, SIGKILL);
